+++ EdSurvey/schedules/templates/newattemptblock.html
@@ -1,0 +1,1 @@
+<a href="{% url 'schedules:newattempt' schedule.id %}">Пройти тест</a>
