+++ EdSurvey/schedules/templates/runattemptblock.html
@@ -1,0 +1,1 @@
+<a href="{% url 'schedules:runattempt' attempt.id %}">Пройти тест</a>
