+++ EdSurvey/schedules/urls.py
@@ -4,4 +4,6 @@ from . import views
 urlpatterns = [
     url(r'^$', views.index),
     url(r'^scheduleinfo/(?P<scheduleid>\d+)$', views.schedule_info, name='scheduleinfo'),
- ]
+    url(r'^runattempt/(?P<attemptid>\d+)$', views.run_attempt, name='runattempt'),
+    url(r'^newattempt/(?P<scheduleid>\d+)$', views.new_attempt, name='newattempt'),
+]
