--- EdSurvey/surveys/templates/newattemptblock.html
@@ -1,1 +1,0 @@
-<a href="{% url 'surveys:newattempt' schedule.id %}">Пройти тест</a>
