--- EdSurvey/surveys/templates/runattemptblock.html
@@ -1,1 +1,0 @@
-<a href="{% url 'surveys:runattempt' attempt.id %}">Пройти тест</a>
