+++ EdSurvey/surveys/urls.py
@@ -6,6 +6,4 @@ urlpatterns = [
     url(r'^choicerun/$', views.choice_run),
     url(r'^choicedone/$', views.choice_done),
     url(r'^choiceattempt/(?P<scheduleid>\d+)$', views.choice_attempt),
-    url(r'^runattempt/(?P<attemptid>\d+)$', views.run_attempt, name='runattempt'),
-    url(r'^newattempt/(?P<scheduleid>\d+)$', views.new_attempt, name='newattempt'),
  ]
