+++ EdSurvey/surveys/views.py
@@ -8,52 +8,6 @@ from schedules.views import render_task_info
 # from querylists.views import render_querylist_info
 
 
-def render_run_attempt(schedule):
-    """ Генерирует HTML-код для отображения возможности запуска попытки пройти тест"""
-    # проверить, что мы в сроках
-    if schedule.start < now() < schedule.finish:
-        # найти незавершённую попытку
-        attempt = Attempt.objects.all().filter(schedule=schedule,
-                                               finished__isnull=True)   # .oreder_by('-started')
-        if attempt:
-            # и вернуть HTML-код запуска теста
-            return render_to_string('runattemptblock.html', {'attempt': attempt[0]})
-        # Если незавершённой попытки нет, то Вычислить количество доступных попыток
-        elif schedule.task.attempts > Attempt.objects.all().filter(schedule=schedule,
-                                                                   finished__isnull=False).count():
-            # Если есть досупные попытки, то вернуть HTML-код запуска теста
-            return render_to_string('newattemptblock.html', {'schedule': schedule})
-        else:
-            # Если все попытки использованы, то сообщить об остутсвие доступных попыток из _имеющихся_
-            return render_to_string('noattemptblock.html', {'attempts': schedule.attempts})
-    else:
-        return render_to_string('outofdateblock.html', {'attempts': schedule.attempts})
-
-
-def run_attempt(request, attemptid):
-    """ Показать общую информацию о параметрах
-    - задания
-    - тест-кейса
-    - конкретной попытки
-    """
-    attempt = get_object_or_404(Attempt, pk=attemptid)
-    return render(
-        request,
-        'runattempt.html',
-        {
-            'attempt': attempt,
-            'taskinfoblock': render_task_info(attempt.schedule.task),
-            # 'querylistinfoblock': render_querylist_info(attempt.schedule.task.querylist),
-        },
-    )
-
-
-def new_attempt(request, scheduleid):
-    attempt = Attempt(schedule=get_object_or_404(Schedule, pk=scheduleid))
-    attempt.save()
-    return run_attempt(request, attempt.id)
-
-
 def index(request):
     """ Показать количество
     - Назначенных аданий        [подробнее >] переход на view выбора задания choice_run с соответсвующим фильтром
